backend/requirements.txt
elevenlabs==0.2.24
Flask==2.3.2
Flask_Cors==4.0.0
openai==0.27.8
python-dotenv==1.0.0
